Allow 64-bit windows and node/nw builds
Previously a change enabled 64-bit builds for nw-gyp; however, this
bit of code was not included in that change.  Removed here.

The nw-gyp application supports building against both nw.lib and
node.lib (what node-gyp builds against).  There's some configuration
for that in this file.  This updates to properly check whether that
case holds.