Add Python2 find script
This script finds Python 2 for compatibility with NW.js.

Unfortunately, node-gyp requires Python 3.  This means that NW and
node-gyp are fundamentally incompatible.  It's not as simple as a
few minor tweaks to convert node-gyp into nw-gyp.  Either node-gyp
needs rewritten to work with Python 2, or NW needs changed to work
with Python 3.  Either approach is a larger project than I can
undertake currently.

This script works.  Perhaps someone else might find it useful.  It
doesn't help with the current project.